+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
  - Send delivery time in seconds rather than milliseconds when tracking delivery
 
+### Fixed
+ - Internal Open Activity to forward the intent extras
+
 ## [0.9.11] - 2018-03-12
 ### Changed
  - Update dependencies to target latest Android API (27) and latest Firebase SDK (11.8.0)